Update blogresults.js
out of curiosity, I tried to make a change on this file with the use of React Hooks. I spent 4 hours trying to get objects, only to result in error of using collection of children instead of an array. I grew frustrated with the madness and I realized that I should use what is the best in this situation and React Hooks is not one of them.